make ExtensionDescriptor an interface
